requirements.txt
# Dependências do Chatbot de Segurança Digital
gradio>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
